backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
Pillow==10.4.0
kagglehub==0.3.4
tensorflow==2.17.0
numpy==1.26.4
